added possibility to generate prefix for data stream node
There might be several objects in a model, which open streams with the
same name. For example, all link objects in a network model might open
streams with the name "bandwidth". In such cases, the data stream name
should be extended with a name of the object, which opened the stream.
Thus, the name of the object (via toString) is used as prefix. The
result are stream names such as "linkA.bandwidth" and "linkB.bandwidth".